[BEAM-6188] Move SyntheticSourceOptions to upper level
